requirements/base.txt
Django==2.0.5
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
Unipath==1.1
pytz==2018.4
django-model-utils==3.1.2
paypalrestsdk==1.13.1
django-countries==5.3.1
